+++ RELEASE_NOTES.md
@@ -1,3 +1,22 @@
+v0.0.5
+------
+Added the ability to pause Inferno after loading a protocol incase the user
+needs to tweak the protocol for that particular cell. This option can be set
+globally via config.ini or on the command line with `-s` or `--stop`. Unless
+`PAUSE ON LOAD = True` is added to the options section of the config or -s is
+set on the command line Iferno will click record immediately after loading the
+protocol.
+
+The OPTIONS section is optional, default values are show in the example config.
+
+#### New config sections
+[OPTIONS]
+PAUSE ON Load = False
+
+#### New command line options
+-s --stop
+
+
 v0.0.4
 ------
 Added the ability to print the meter values for MultiClampCommander.